requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
langgraph==0.0.30
langchain-core==0.1.33
langchain-anthropic==0.1.4
tavily-python==0.3.3
qdrant-client==1.8.1
voyageai==0.2.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.1.1
